HP Pavilion dv6.xml
I fixed major & minor issues with the file you provided users of the Pavilion dv6.
First the critical temp is now 90C, this is because this laptop shuts down when reaching >100C.
Second I changed the temp. threshold!
Lastly I changed the fan speed so the laptop fan would be able to run at its max.