Create HP Laptop 17-ca1xxx.xml
This config has been created manually using the ec probing tool and tested on an HP 17-ca1211ng Jet Black (7KG31EA#ABD).
The configuration is based on the "HP Laptop 14-cm0xxx.xml" from  Matias Hiltunen and the description was partially taken from Zwergesel (HP 17-by1102ng) - thanks to both.

Remark: It had to disable the BIOS option "System Configuration" > "Fan Always On" in order to make the fan turn off completely.

The configuration uses the write register 219/0xDB in a range from 34 to 80 and the read register 113/0x71 in a range from 0 to 7 based on educated guessing, so there could be better ones.
The fan seems to have 8 levels (including 0) and the read register maps nicely. 
The write register mapping is a bit exponential towards the end. And having 55 distinct values mapping to 8 stages seems also a bit confusing, e.g. when using the scroll bar in the UI.

The temperature thresholds have been mapped for my convenience so that the fan starts somewhere at 55 degrees, therefore feel free to change as you like.
The ranges go up 80 degrees with some overlap, and the critical temperatur is also at 80 degrees.
This is based on my experience that the AMD Ryzen 3 3200U hardly crossed 65 degrees, with or without fan, even when the spec allows up to 105 degrees. Which means there might be some additional guards build-in anyway.

And I have to say thanks for this fantastic tool. Without the fan never stopped and I was close to send the new notebook back to HP straight ahead.