Added profile for HP ZBook Studio x360
This is a profile for HP ZBook Studio x360 that I created based on the default HP ZBook Studio G3 profile. The two models share the same EC registers but with different Min. Speed and Max. Speed values. I used RWE to confirm the correct maximum fan speed value that the EC accepted, and reflected my findings in the profile.
The Min. Speed value currently doesn't correspond to 0 RPM, which requires further investigation.